+++ noodles-util/src/variant/io/indexed_reader.rs
@@ -73,7 +73,7 @@ where
             Self::Vcf(reader) => Box::new(
                 reader
                     .query(header, region)?
-                    .into_iter()
+                    .records()
                     .map(|result| result.map(|record| Box::new(record) as Box<dyn Record>)),
             ),
             Self::Bcf(reader) => Box::new(