+++ noodles-vcf/src/io/reader/query.rs
@@ -10,9 +10,9 @@ use self::records::Records;
 use super::Reader;
 use crate::{Header, Record, variant::Record as _};
 
-/// An iterator over records of a VCF reader that intersects a given region.
+/// A reader over records of a VCF reader that intersects a given region.
 ///
-/// This is created by calling [`super::Reader::query`].
+/// This is created by calling [`Reader::query`].
 pub struct Query<'r, 'h, R> {
     reader: Reader<csi::io::Query<'r, R>>,
     header: &'h Header,
@@ -39,7 +39,8 @@ where
         }
     }
 
-    fn read_record(&mut self, record: &mut Record) -> io::Result<usize> {
+    /// Reads a record.
+    pub fn read_record(&mut self, record: &mut Record) -> io::Result<usize> {
         next_record(
             &mut self.reader,
             record,
@@ -48,16 +49,9 @@ where
             self.interval,
         )
     }
-}
 
-impl<'r, 'h: 'r, R> IntoIterator for Query<'r, 'h, R>
-where
-    R: bgzf::io::BufRead + bgzf::io::Seek,
-{
-    type Item = io::Result<Record>;
-    type IntoIter = Records<'r, 'h, R>;
-
-    fn into_iter(self) -> Self::IntoIter {
+    /// Returns an iterator over records.
+    pub fn records(self) -> Records<'r, 'h, R> {
         Records::new(self)
     }
 }